+++ src/Bundles/Translator/TranslatorFactory.php
@@ -3,16 +3,7 @@
 namespace PHPUnuhi\Bundles\Translator;
 
 use Exception;
-use PHPUnuhi\Bundles\Storage\INI\IniStorage;
-use PHPUnuhi\Bundles\Storage\JSON\JsonStorage;
-use PHPUnuhi\Bundles\Storage\PHP\PhpStorage;
-use PHPUnuhi\Bundles\Storage\PO\PoStorage;
-use PHPUnuhi\Bundles\Storage\Shopware6\Shopware6Storage;
-use PHPUnuhi\Bundles\Storage\StorageFactory;
-use PHPUnuhi\Bundles\Storage\StorageInterface;
-use PHPUnuhi\Bundles\Storage\YAML\YamlStorage;
 use PHPUnuhi\Bundles\Translator\DeepL\DeeplTranslator;
-use PHPUnuhi\Bundles\Translator\Fake\FakeTranslator;
 use PHPUnuhi\Bundles\Translator\GoogleCloud\GoogleCloudTranslator;
 use PHPUnuhi\Bundles\Translator\GoogleWeb\GoogleWebTranslator;
 use PHPUnuhi\Bundles\Translator\OpenAI\OpenAITranslator;
@@ -82,7 +73,6 @@ class TranslatorFactory
     {
         $this->translators = [];
 
-        $this->translators[] = new FakeTranslator();
         $this->translators[] = new DeeplTranslator();
         $this->translators[] = new OpenAITranslator();
         $this->translators[] = new GoogleWebTranslator();