+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes of PHPUnuhi releases are documented in this file
 using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
+## [unreleased]
+
+### Removed
+
+- Removed **fake** translator service. This was accidentally existed inside the code and factory, but was only meant for
+  unit tests.
+
 ## [1.14.0]
 
 ### Added